tests/wpt/mozilla/tests/mozilla/proxy_setter.html

<html>
<head>
  <title></title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
test(function() {
  assert_equals(window.document.title, '');
  window.document.title = 'foo';
  assert_equals(window.document.title, 'foo');
});
</script>
</body>
</html>
